+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
   * New action added: :class:`mockify.actions.YieldAsync`
   * New action added: :class:`mockify.actions.ReturnContext`
+  * New action added: :class:`mockify.actions.ReturnAsyncContext`
 
 0.11.0 (2020-11-24)
 -------------------